+++ Core.lua
@@ -24,10 +24,6 @@ function BLU:OnInitialize()
 	self:RegisterChatCommand("blu", "SlashCommand")
 end
 
-
-
-
-
 function BLU:OnEnable()
 	self:RegisterEvent("PLAYER_STARTED_MOVING")
 	self:RegisterEvent("PLAYER_LEVEL_UP")
@@ -37,63 +33,53 @@ function BLU:PLAYER_LEVEL_UP(self, event, ...)
 	if BLU.db.profile.LevelSoundSelect == 1 then
 	PlaySoundFile(569593)
 elseif BLU.db.profile.LevelSoundSelect == 2 then
-	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\FFLU.ogg", "Master")
+	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\FFLU.ogg", "SFX")
 elseif BLU.db.profile.LevelSoundSelect == 3 then
-	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\FNLU.ogg", "Master")
+	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\FNLU.ogg", "SFX")
 elseif BLU.db.profile.LevelSoundSelect == 4 then
-	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\KH3LU.ogg", "Master")
+	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\KH3LU.ogg", "SFX")
 elseif BLU.db.profile.LevelSoundSelect == 5 then
-	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\LoLLU.ogg", "Master")
+	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\LoLLU.ogg", "SFX")
 elseif BLU.db.profile.LevelSoundSelect == 6 then
-	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\MSLU.ogg", "Master")
+	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\MSLU.ogg", "SFX")
 elseif BLU.db.profile.LevelSoundSelect == 7 then
-	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\MCLU.ogg", "Master")
+	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\MCLU.ogg", "SFX")
 elseif BLU.db.profile.LevelSoundSelect == 8 then
-	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\MW2LU.ogg", "Master")
+	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\MW2LU.ogg", "SFX")
 elseif BLU.db.profile.LevelSoundSelect == 9 then
-	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\MWLU.ogg", "Master")
+	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\MWLU.ogg", "SFX")
 elseif BLU.db.profile.LevelSoundSelect == 10 then
-	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\OSRSLU.ogg", "Master")
+	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\OSRSLU.ogg", "SFX")
 elseif BLU.db.profile.LevelSoundSelect == 11 then
-	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\PoELU.ogg", "Master")
+	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\PoELU.ogg", "SFX")
 elseif BLU.db.profile.LevelSoundSelect == 12 then
-	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\PkmnLU.ogg", "Master")
+	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\PkmnLU.ogg", "SFX")
 elseif BLU.db.profile.LevelSoundSelect == 13 then
-	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\SRLU.ogg", "Master")
+	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\SRLU.ogg", "SFX")
 elseif BLU.db.profile.LevelSoundSelect == 14 then
-	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\SHHLU.ogg", "Master")
+	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\SHHLU.ogg", "SFX")
 elseif BLU.db.profile.LevelSoundSelect == 15 then
-	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\SMB3LU.ogg", "Master")
+	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\SMB3LU.ogg", "SFX")
 elseif BLU.db.profile.LevelSoundSelect == 16 then
-	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\WC3LU.ogg", "Master")
+	PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\WC3LU.ogg", "SFX")
 end
 end
-
 
 function BLU:PLAYER_STARTED_MOVING(event)
-	if BLU.db.profile.MuteLevelDefault == true then
+	if BLU.db.profile.MuteLevelDefault == true and BLU.db.profile.MuteRepDefault == true then
 	MuteSoundFile(569593)
-elseif BLU.db.profile.MuteLevelDefault == false then
-	UnmuteSoundFile(569593)
-end
-end
-
-
-
-
-
-function BLU:PLAYER_STARTED_MOVING(event)
-	if BLU.db.profile.MuteRepDefault == true then
 	MuteSoundFile(568016)
-elseif BLU.db.profile.MuteRepDefault == false then
+elseif BLU.db.profile.MuteLevelDefault == false and BLU.db.profile.MuteRepDefault == false then
+	UnmuteSoundFile(569593)
 	UnmuteSoundFile(568016)
+elseif BLU.db.profile.MuteLevelDefault == true and BLU.db.profile.MuteRepDefault == false then
+	MuteSoundFile(569593)
+	UnmuteSoundFile(568016)
+elseif BLU.db.profile.MuteLevelDefault == false and BLU.db.profile.MuteRepDefault == true then
+	UnmuteSoundFile(569593)
+	MuteSoundFile(568016)
 end
 end
-
-
-
-
-
 
 local TrackedFactions={};
 local Event_Frame_B=CreateFrame("Frame");
@@ -106,49 +92,44 @@ Event_Frame_B:SetScript("OnEvent",function()
 	       local _,_,newstanding,_,_,_,_,_,isheader,_,hasrep,_,_,faction=GetFactionInfo(i);
 	       if faction and (not isheader or hasrep) and (newstanding or 0)>0 then-- Make sure we have the info we need and standing isn't UNKNOWN (0)
 	           local oldstanding=TrackedFactions[faction];
-						 if oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 1 then-- Check if standing went up (allow same code to initialize tracking)
+						 if oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 1 then
 	                PlaySoundFile(568016)
-	           elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 2 then-- Check if standing went up (allow same code to initialize tracking)
-	                PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\FFLU.ogg", "Master")
-								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 3 then-- Check if standing went up (allow same code to initialize tracking)
-			             PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\FNLU.ogg", "Master")
-								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 4 then-- Check if standing went up (allow same code to initialize tracking)
-									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\KH3LU.ogg", "Master")
-								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 5 then-- Check if standing went up (allow same code to initialize tracking)
-									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\LoLLU.ogg", "Master")
-								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 6 then-- Check if standing went up (allow same code to initialize tracking)
-									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\MSLU.ogg", "Master")
-								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 7 then-- Check if standing went up (allow same code to initialize tracking)
-									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\MCLU.ogg", "Master")
-								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 8 then-- Check if standing went up (allow same code to initialize tracking)
-									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\MW2LU.ogg", "Master")
-								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 9 then-- Check if standing went up (allow same code to initialize tracking)
-									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\MWLU.ogg", "Master")
-								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 10 then-- Check if standing went up (allow same code to initialize tracking)
-									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\OSRSLU.ogg", "Master")
-								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 11 then-- Check if standing went up (allow same code to initialize tracking)
-									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\PoELU.ogg", "Master")
-								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 12 then-- Check if standing went up (allow same code to initialize tracking)
-									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\PkmnLU.ogg", "Master")
-								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 13 then-- Check if standing went up (allow same code to initialize tracking)
-									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\SRLU.ogg", "Master")
-								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 14 then-- Check if standing went up (allow same code to initialize tracking)
-									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\SHHLU.ogg", "Master")
-								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 15 then-- Check if standing went up (allow same code to initialize tracking)
-									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\SMB3LU.ogg", "Master")
-								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 16 then-- Check if standing went up (allow same code to initialize tracking)
-									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\WC3LU.ogg", "Master")
+	           elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 2 then
+	                PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\FFLU.ogg", "SFX")
+								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 3 then
+			             PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\FNLU.ogg", "SFX")
+								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 4 then
+									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\KH3LU.ogg", "SFX")
+								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 5 then
+									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\LoLLU.ogg", "SFX")
+								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 6 then
+									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\MSLU.ogg", "SFX")
+								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 7 then
+									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\MCLU.ogg", "SFX")
+								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 8 then
+									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\MW2LU.ogg", "SFX")
+								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 9 then
+									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\MWLU.ogg", "SFX")
+								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 10 then
+									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\OSRSLU.ogg", "SFX")
+								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 11 then
+									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\PoELU.ogg", "SFX")
+								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 12 then
+									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\PkmnLU.ogg", "SFX")
+								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 13 then
+									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\SRLU.ogg", "SFX")
+								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 14 then
+									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\SHHLU.ogg", "SFX")
+								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 15 then
+									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\SMB3LU.ogg", "SFX")
+								elseif oldstanding and oldstanding<newstanding and BLU.db.profile.RepSoundSelect == 16 then
+									PlaySoundFile("Interface\\Addons\\BLU\\Sounds\\WC3LU.ogg", "SFX")
 	           end
 
 	           TrackedFactions[faction]=newstanding;-- Update standing
 	       end
 	   end
 end);
-
-
-
-
-
 
 function BLU:SlashCommand(input, editbox)
 
@@ -159,12 +140,7 @@ function BLU:SlashCommand(input, editbox)
 			-- unregisters all events and calls BLU:OnDisable() if you defined that
 			self:Disable()
 			self:Print("Disabled.")
-		else--[[if input == "message" then
-			print("this is our saved message:", self.db.profile.someInput)
-		else]]
-
-
-
+		else
 
 		-- https://github.com/Stanzilla/WoWUIBugs/issues/89
 		InterfaceOptionsFrame_OpenToCategory(self.optionsFrame)